Rewrite the website for hosting on Firebase
After much soulsearching (and learning about Node), I'm porting my website to Firebase hosting.
It'll still be free for me (yay) but it requires a slightly different app structure. I've removed the hosting code to let them handle it quicker, and I've included React for a rewrite of the site I'm going to do sometime soon.